+++ packages/at_auth/lib/at_auth.dart
@@ -2,9 +2,17 @@ library at_auth;
 
 export 'src/at_auth_base.dart';
 export 'src/at_auth_impl.dart';
+export 'src/auth_constants.dart';
 export 'src/onboard/at_onboarding_request.dart';
 export 'src/onboard/at_onboarding_response.dart';
 export 'src/auth/at_auth_request.dart';
 export 'src/auth/at_auth_response.dart';
 export 'src/keys/at_auth_keys.dart';
 export 'src/exception/at_auth_exceptions.dart';
+export 'src/enroll/at_enrollment_base.dart';
+export 'src/enroll/at_enrollment_impl.dart';
+export 'src/enroll/at_enrollment_response.dart';
+export 'src/enroll/at_enrollment_request.dart';
+export 'src/enroll/at_initial_enrollment_request.dart';
+export 'src/enroll/at_new_enrollment_request.dart';
+export 'src/enroll/at_enrollment_notification_request.dart';