+++ packages/at_auth/lib/src/at_auth_impl.dart
@@ -4,6 +4,10 @@ import 'dart:io';
 import 'package:at_auth/src/at_auth_base.dart';
 import 'package:at_auth/src/auth/cram_authenticator.dart';
 import 'package:at_auth/src/auth/pkam_authenticator.dart';
+import 'package:at_auth/src/enroll/at_enrollment_base.dart';
+import 'package:at_auth/src/enroll/at_enrollment_impl.dart';
+import 'package:at_auth/src/enroll/at_enrollment_response.dart';
+import 'package:at_auth/src/enroll/at_initial_enrollment_request.dart';
 import 'package:at_auth/src/keys/at_auth_keys.dart';
 import 'package:at_auth/src/onboard/at_onboarding_request.dart';
 import 'package:at_auth/src/onboard/at_onboarding_response.dart';
@@ -26,18 +30,22 @@ class AtAuthImpl implements AtAuth {
 
   PkamAuthenticator? pkamAuthenticator;
 
+  AtEnrollmentBase? atEnrollmentBase;
+
   AtLookUp? atLookUp;
 
   AtAuthImpl(
       {this.atLookUp,
       this.atChops,
       this.cramAuthenticator,
-      this.pkamAuthenticator});
+      this.pkamAuthenticator,
+      this.atEnrollmentBase});
 
   @override
   Future<AtAuthResponse> authenticate(AtAuthRequest atAuthRequest) async {
     if (atAuthRequest.atKeysFilePath == null &&
-        atAuthRequest.atAuthKeys == null) {
+        atAuthRequest.atAuthKeys == null &&
+        atAuthRequest.encryptedKeysMap == null) {
       throw AtAuthenticationException(
           'Keyfile path or atAuthKeys object has to be set in atAuthRequest');
     }
@@ -49,14 +57,18 @@ class AtAuthImpl implements AtAuth {
       atAuthKeys = _decryptAtKeysFile(
           await _readAtKeysFile(atAuthRequest.atKeysFilePath),
           atAuthRequest.authMode);
-      enrollmentIdFromRequest ??= atAuthKeys.enrollmentId;
+    } else if (atAuthRequest.encryptedKeysMap != null) {
+      atAuthKeys = _decryptAtKeysFile(
+          atAuthRequest.encryptedKeysMap!, PkamAuthMode.keysFile);
     } else {
       atAuthKeys = atAuthRequest.atAuthKeys;
     }
     if (atAuthKeys == null) {
       throw AtAuthenticationException(
-          'keys either were not provided in the AtAuthRequest, or could not be read from provided keys file');
+          'keys either were not provided in the AtAuthRequest,'
+          ' or could not be read from provided keys file');
     }
+    enrollmentIdFromRequest ??= atAuthKeys.enrollmentId;
     var pkamPrivateKey = atAuthKeys.apkamPrivateKey;
 
     if (atAuthRequest.authMode == PkamAuthMode.keysFile &&
@@ -77,16 +89,19 @@ class AtAuthImpl implements AtAuth {
       var pkamResponse = (await pkamAuthenticator!
           .authenticate(enrollmentId: enrollmentIdFromRequest));
       isPkamAuthenticated = pkamResponse.isSuccessful;
-    } on Exception catch (e) {
-      _logger.severe('Caught exception: $e');
+    } on AtException catch (e) {
+      _logger.severe('Caught $e');
       throw AtAuthenticationException(
-          'Unable to authenticate- ${e.toString()}');
+          'Unable to authenticate | Cause: ${e.message}');
+    } on Exception catch (e) {
+      throw AtAuthenticationException('Unable to authenticate | Cause: $e');
     }
     _logger.finer(
         'PKAM auth result: ${isPkamAuthenticated ? 'success' : 'failed'}');
     return AtAuthResponse(atAuthRequest.atSign)
       ..isSuccessful = isPkamAuthenticated
-      ..enrollmentId = enrollmentIdFromRequest;
+      ..enrollmentId = enrollmentIdFromRequest
+      ..atAuthKeys = atAuthKeys;
   }
 
   @override
@@ -103,7 +118,7 @@ class AtAuthImpl implements AtAuth {
     if (!cramAuthResult.isSuccessful) {
       throw AtAuthenticationException(
           'Cram authentication failed. Please check the cram key'
-          ' and try again \n(or) contact support@atsign.com');
+          ' and try again (or) contact support@atsign.com');
     }
     //2. generate key pairs
     var atAuthKeys = _generateKeyPairs(atOnboardingRequest.authMode,
@@ -117,7 +132,8 @@ class AtAuthImpl implements AtAuth {
     //3. update pkam public key through enrollment or manually based on app preference
     String? enrollmentIdFromServer;
     if (atOnboardingRequest.enableEnrollment) {
-      // server will update the apkam public key during enrollment.So don't have to manually update in this scenario.
+      // server will update the apkam public key during enrollment.
+      // So don't have to manually update in this scenario.
       enrollmentIdFromServer = await _sendOnboardingEnrollment(
           atOnboardingRequest, atAuthKeys, atLookUp!);
       atAuthKeys.enrollmentId = enrollmentIdFromServer;
@@ -132,20 +148,20 @@ class AtAuthImpl implements AtAuth {
       _logger.finer('PkamPublicKey update result: $pkamUpdateResult');
     }
 
-    //3. Close connection to server
+    //4. Close connection to server
     try {
       await (atLookUp as AtLookupImpl).close();
     } on Exception catch (e) {
       _logger.severe('error while closing connection to server: $e');
     }
 
-    //4. Init _atLookUp again and attempt pkam auth
+    //5. Init _atLookUp again and attempt pkam auth
     // atLookUp = AtLookupImpl(atOnboardingRequest.atSign,
     //     atOnboardingRequest.rootDomain, atOnboardingRequest.rootPort);
     atLookUp!.atChops = atChops;
 
     var isPkamAuthenticated = false;
-    //5. Do pkam auth
+    //6. Do pkam auth
     pkamAuthenticator ??=
         PkamAuthenticator(atOnboardingRequest.atSign, atLookUp!);
     try {
@@ -159,7 +175,8 @@ class AtAuthImpl implements AtAuth {
       throw AtAuthenticationException('Pkam auth returned false');
     }
 
-    //5. If Pkam auth is success, update encryption public key to secondary and delete cram key from server
+    //7. If Pkam auth is success, update encryption public key to secondary
+    // and delete cram key from server
     final encryptionPublicKey = atAuthKeys.defaultEncryptionPublicKey;
     UpdateVerbBuilder updateBuilder = UpdateVerbBuilder()
       ..atKey = 'publickey'
@@ -168,7 +185,8 @@ class AtAuthImpl implements AtAuth {
       ..sharedBy = atOnboardingRequest.atSign;
     String? encryptKeyUpdateResult = await atLookUp!.executeVerb(updateBuilder);
     _logger.info('Encryption public key update result $encryptKeyUpdateResult');
-    // deleting cram secret from the keystore as cram auth is complete
+
+    //8.  Delete cram secret from the keystore as cram auth is complete
     DeleteVerbBuilder deleteBuilder = DeleteVerbBuilder()
       ..atKey = AtConstants.atCramSecret;
     String? deleteResponse = await atLookUp!.executeVerb(deleteBuilder);
@@ -198,40 +216,42 @@ class AtAuthImpl implements AtAuth {
       AtOnboardingRequest atOnboardingRequest,
       AtAuthKeys atAuthKeys,
       AtLookUp atLookup) async {
-    var enrollBuilder = EnrollVerbBuilder()
-      ..appName = atOnboardingRequest.appName
-      ..deviceName = atOnboardingRequest.deviceName;
-
     var symmetricEncryptionAlgo =
         AESEncryptionAlgo(AESKey(atAuthKeys.apkamSymmetricKey!));
-    enrollBuilder.encryptedDefaultEncryptedPrivateKey = atChops!
+    var encryptedDefaultEncryptionPrivateKey = atChops!
         .encryptString(
             atAuthKeys.defaultEncryptionPrivateKey!, EncryptionKeyType.aes256,
             encryptionAlgorithm: symmetricEncryptionAlgo,
             iv: AtChopsUtil.generateIVLegacy())
         .result;
-
-    enrollBuilder.encryptedDefaultSelfEncryptionKey = atChops!
+    var encryptedDefaultSelfEncryptionKey = atChops!
         .encryptString(
             atAuthKeys.defaultSelfEncryptionKey!, EncryptionKeyType.aes256,
             encryptionAlgorithm: symmetricEncryptionAlgo,
             iv: AtChopsUtil.generateIVLegacy())
         .result;
-    enrollBuilder.apkamPublicKey = atAuthKeys.apkamPublicKey;
-
-    var enrollResult = await atLookup
-        .executeCommand(enrollBuilder.buildCommand(), auth: false);
-    if (enrollResult == null || enrollResult.isEmpty) {
-      throw AtAuthenticationException('Enrollment response is null or empty');
-    } else if (enrollResult.startsWith('error:')) {
-      throw AtAuthenticationException('Enrollment error:$enrollResult');
+    _logger.finer('apkamPublicKey: ${atAuthKeys.apkamPublicKey}');
+    var enrollRequestBuilder = AtInitialEnrollmentRequestBuilder()
+      ..setAppName(atOnboardingRequest.appName)
+      ..setDeviceName(atOnboardingRequest.deviceName)
+      ..setEncryptedDefaultEncryptionPrivateKey(
+          encryptedDefaultEncryptionPrivateKey)
+      ..setEncryptedDefaultSelfEncryptionKey(encryptedDefaultSelfEncryptionKey)
+      ..setApkamPublicKey(atAuthKeys.apkamPublicKey)
+      ..setAtAuthKeys(atAuthKeys)
+      ..setEnrollOperationEnum(EnrollOperationEnum.request);
+    atEnrollmentBase ??= AtEnrollmentImpl(atOnboardingRequest.atSign);
+    AtEnrollmentResponse enrollmentResponse;
+    try {
+      enrollmentResponse = await atEnrollmentBase!
+          .submitEnrollment(enrollRequestBuilder.build(), atLookUp!);
+    } on AtEnrollmentException catch (e) {
+      throw AtAuthenticationException('Enrollment error:${e.toString}');
     }
-    enrollResult = enrollResult.replaceFirst('data:', '');
-    _logger.finer('enrollResult: $enrollResult');
-    var enrollResultJson = jsonDecode(enrollResult);
-    var enrollmentIdFromServer = enrollResultJson[AtConstants.enrollmentId];
-    var enrollmentStatus = enrollResultJson['status'];
-    if (enrollmentStatus != 'approved') {
+    _logger.finer('enrollment response: ${enrollmentResponse.toString()}');
+    var enrollmentIdFromServer = enrollmentResponse.enrollmentId;
+    var enrollmentStatus = enrollmentResponse.enrollStatus;
+    if (enrollmentStatus != EnrollStatus.approved) {
       throw AtAuthenticationException(
           'initial enrollment is not approved. Status from server: $enrollmentStatus');
     }
@@ -239,7 +259,7 @@ class AtAuthImpl implements AtAuth {
   }
 
   AtAuthKeys _decryptAtKeysFile(
-      Map<String, String> jsonData, PkamAuthMode authMode) {
+      Map<String, dynamic> jsonData, PkamAuthMode authMode) {
     var securityKeys = AtAuthKeys();
     String decryptionKey = jsonData[auth_constants.defaultSelfEncryptionKey]!;
     var atChops =