+++ packages/at_commons/lib/src/at_constants.dart
@@ -1,5 +1,3 @@
-part 'at_constants_legacy.dart';
-
 class AtConstants {
   static const String atSign = 'atSign';
   static const String forAtSign = 'forAtSign';
@@ -20,6 +18,8 @@ class AtConstants {
   static const String atSigningPublicKey = 'public:signing_publickey';
   static const String atCramSecret = 'privatekey:at_secret';
   static const String atCramSecretDeleted = 'privatekey:at_secret_deleted';
+  static const String atBlocklist =
+      'private:blocklist'; // contains @atsign postfix
   static const String atSigningKeypairGenerated =
       'privatekey:signing_keypair_generated';
   static const String statId = 'statId';
@@ -28,7 +28,6 @@ class AtConstants {
   static const String ttr = 'ttr';
   static const String ttlNotification = 'ttln';
   static const String fromCommitSequence = 'from_commit_seq';
-  static const String atOperation = 'operation';
   static const String atValueReference = 'atsign://';
   static const String regex = 'regex';
   static const String configNew = 'configNew';