+++ packages/at_lookup/lib/src/at_lookup_impl.dart
@@ -5,6 +5,7 @@ import 'dart:convert';
 import 'dart:io';
 import 'dart:typed_data';
 
+import 'package:at_chops/at_chops.dart';
 import 'package:at_commons/at_builders.dart';
 import 'package:at_commons/at_commons.dart';
 import 'package:at_lookup/at_lookup.dart';
@@ -13,7 +14,6 @@ import 'package:at_utils/at_logger.dart';
 import 'package:crypto/crypto.dart';
 import 'package:crypton/crypton.dart';
 import 'package:mutex/mutex.dart';
-import 'package:at_chops/at_chops.dart';
 
 class AtLookupImpl implements AtLookUp {
   final logger = AtSignLogger('AtLookup');
@@ -34,6 +34,7 @@ class AtLookupImpl implements AtLookUp {
 
   late int _rootPort;
 
+  @Deprecated("use atChops")
   String? privateKey;
 
   String? cramSecret;
@@ -333,7 +334,7 @@ class AtLookupImpl implements AtLookUp {
 
   Future<String> _update(UpdateVerbBuilder builder) async {
     String atCommand;
-    if (builder.operation == UPDATE_META) {
+    if (builder.operation == AtConstants.updateMeta) {
       atCommand = builder.buildCommandForMeta();
     } else {
       atCommand = builder.buildCommand();
@@ -424,7 +425,7 @@ class AtLookupImpl implements AtLookUp {
         logger.finer('fromResponse $fromResponse');
         var key = RSAPrivateKey.fromString(privateKey);
         var sha256signature =
-            key.createSHA256Signature(utf8.encode(fromResponse) as Uint8List);
+             key.createSHA256Signature(utf8.encode(fromResponse) as Uint8List);
         var signature = base64Encode(sha256signature);
         logger.finer('Sending command pkam:$signature');
         await _sendCommand('pkam:$signature\n');
@@ -492,14 +493,8 @@ class AtLookupImpl implements AtLookUp {
 
   final Mutex _cramAuthenticationMutex = Mutex();
 
-  /// Generates digest using from verb response and [secret] and performs a CRAM authentication to
-  /// secondary server
-  // ignore: non_constant_identifier_names
-  Future<bool> authenticate_cram(var secret) async {
-    secret ??= cramSecret;
-    if (secret == null) {
-      throw UnAuthenticatedException('Cram secret not passed');
-    }
+  @override
+  Future<bool> cramAuthenticate(String secret) async {
     await createConnection();
     try {
       await _cramAuthenticationMutex.acquire();
@@ -532,6 +527,16 @@ class AtLookupImpl implements AtLookUp {
     } finally {
       _cramAuthenticationMutex.release();
     }
+  }
+
+  @Deprecated('use AtLookup().cramAuthenticate()')
+  // ignore: non_constant_identifier_names
+  Future<bool> authenticate_cram(var secret) async {
+    secret ??= cramSecret;
+    if (secret == null) {
+      throw UnAuthenticatedException('Cram secret not passed');
+    }
+    return await cramAuthenticate(secret);
   }
 
   Future<String> _plookup(PLookupVerbBuilder builder) async {
@@ -573,7 +578,7 @@ class AtLookupImpl implements AtLookUp {
           logger.finer('calling pkam without atchops');
           await authenticate(privateKey);
         } else if (cramSecret != null) {
-          await authenticate_cram(cramSecret);
+          await cramAuthenticate(cramSecret!);
         } else {
           throw UnAuthenticatedException(
               'Unable to perform atLookup auth. Private key/cram secret is not set');
@@ -622,6 +627,7 @@ class AtLookupImpl implements AtLookUp {
     return _connection!.isInValid();
   }
 
+  @override
   Future<void> close() async {
     await _connection!.close();
   }